Code cleanup + Maven
Added gitignore
Converted to Maven project
Minor code cleanup + changes
Version bump to 2.0.0 for semantic versioning - major overhaul incoming!
